More work on Moss Dog
Did some more work on Moss Dog. Can't guarantee it all works perfectly, but I made some adjustments to the Character script to help with Moss Dog along with commenting on those areas for future reference. Bassed some of the code on the Frog code, but can't be sure if they work (or make sense).